Refactor start script into modules
